chore: cleanup files and tests
