+++ packages/engine/src/step-handler.registry.ts
@@ -1,10 +1,8 @@
-// import { ActionStepHandler } from "./step-handlers/action.handler.js";
 import type { ResolveStepArgs } from "./resolve.js";
 import { PipeResolver } from "./pipe-resolver.js";
 import { McpStepHandler } from "./step-handlers/mcp.handler.js";
 
 export type StepHandlerRegistry = {
-  // action: ActionStepHandler;
   mcp: McpStepHandler;
 };
 
@@ -13,7 +11,6 @@ export function wireStepHandlers(
   pipeResolver: PipeResolver
 ): StepHandlerRegistry {
   const stepHandlers = {
-    // action: new ActionStepHandler(argResolver, pipeResolver),
     mcp: new McpStepHandler(argResolver, pipeResolver),
   };
   return stepHandlers;
