+++ packages/events/src/emitters/base.emitter.ts
@@ -2,9 +2,24 @@ import { CloudScope } from "@pipewarp/types";
 import { OtelContext, EnvelopeHeader } from "../types.js";
 
 /**
- * base emitter class all other emitters inherit.
- * currently helps create base envelope fields from otel context
- * and cloud scopes
+ * NOTE: This current design of the emitters has a few loose ends:
+ *
+ * Context creation is in the process of being unified as more detail
+ * about the creation site for these emitters is gathered.  The class is
+ * currently in between the process of moving scope from the EmitterFactory
+ * to the emitter classes.
+ *
+ * Questions remain about how flexible emitters should be.
+ *
+ * OTEL creation is currently too manual and should not be held by
+ * EmitterFactory contexts.
+ *
+ * A separate Otel class or their SDK should be used to create complex
+ * nested parent span relationships.
+ *
+ * Overall the EmitterFactory and BaseEmitter or other emitters need to
+ * be refactored for better DX and Otel features.
+ *
  */
 export class BaseEmitter {
   protected otel: OtelContext;
