+++ packages/events/tests/step-emitter.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from "vitest";
-import { StepEmitter } from "../src/step-emitter.js";
+import { StepEmitter } from "../src/emitters/step.emitter.js";
 import { EventBusPort } from "@pipewarp/ports";
 import { AnyEvent } from "@pipewarp/types";
 import { afterEach } from "node:test";
@@ -20,54 +20,52 @@ describe("[step-emitter]", () => {
     vi.spyOn(crypto, "randomUUID").mockReturnValue(testId as any);
     vi.spyOn(Date.prototype, "toISOString").mockReturnValue(testTime);
 
-    const event: AnyEvent<"step.action.completed"> = {
+    const event: AnyEvent<"step.started"> = {
       spanid: "test-spanid",
       traceid: "test-traceid",
       traceparent: "test-traceparent",
-      parentspanid: "test-parentspanid",
       flowid: "test-flowid",
       runid: "test-runid",
       stepid: "test-stepid",
+      steptype: "test-steptype",
       source: "test-source",
       domain: "step",
-      entity: "action",
-      action: "completed",
+      action: "started",
       data: {
-        ok: false,
-        message: "test-message",
+        status: "started",
+        step: {
+          id: testId,
+          name: "test-name",
+          type: "test-steptype",
+        },
       },
       specversion: "1.0",
-      type: "step.action.completed",
+      type: "step.started",
       id: testId,
       time: testTime,
     };
-    const emitter = new StepEmitter(
-      bus,
-      {
-        spanId: event.spanid,
-        traceId: event.traceid,
-        traceParent: event.traceparent,
-        parentSpanId: event.parentspanid,
-      },
-      {
-        flowid: event.flowid,
-        runid: event.runid,
-        stepid: event.stepid,
-      },
-      {
-        source: event.source,
-      }
-    );
+    const emitter = new StepEmitter(bus, {
+      spanId: event.spanid,
+      traceId: event.traceid,
+      traceParent: event.traceparent,
+      parentSpanId: event.parentspanid,
+      flowid: event.flowid,
+      runid: event.runid,
+      stepid: event.stepid,
+      source: event.source,
+      steptype: event.steptype,
+    });
 
-    emitter.emit("step.action.completed", {
-      ok: event.data.ok,
-      message: event.data.message,
+    emitter.emit("step.started", {
+      status: "started",
+      step: {
+        id: testId,
+        name: "test-name",
+        type: "test-steptype",
+      },
     });
 
     expect(publish).toHaveBeenCalledOnce();
-    expect(publish).toHaveBeenCalledWith(
-      registry["step.action.completed"].topic,
-      event
-    );
+    expect(publish).toHaveBeenCalledWith(registry["step.started"].topic, event);
   });
 });
