+++ packages/types/src/events/engine/event.ts
@@ -1,6 +1,9 @@
 import { CloudEvent } from "../shared/cloud-event.js";
-import type { EngineScope } from "./scope.js";
 import type { EngineEventType } from "./map.js";
+
+export type EngineScope = {
+  engineid: string;
+};
 
 export type EngineEvent<T extends EngineEventType> = EngineScope &
   CloudEvent<T>;
