--- packages/types/src/events/engine/scope.ts
@@ -1,3 +1,0 @@
-export type EngineScope = {
-  engineid: string;
-};
