+++ packages/types/src/events/flow/data.ts
@@ -1,5 +1,3 @@
-import type { FlowEvent } from "../shared/flow-event.js";
-
 export type FlowDescriptor = {
   flow: {
     id: string;
@@ -17,5 +15,3 @@ export type FlowQueuedData = FlowDescriptor & {
 
 export type FlowStartedData = FlowDescriptor & {};
 export type FlowCompletedData = FlowDescriptor & {};
-
-export type FlowQueuedEvent = FlowEvent<"flow.queued">;
