+++ packages/types/src/events/run/index.ts
@@ -1,4 +1,3 @@
 export * from "./data.js";
-export * from "./generic.js";
+export * from "./event.js";
 export * from "./map.js";
-export * from "./scope.js";
