--- packages/types/src/events/run/scope.ts
@@ -1,4 +1,0 @@
-export type RunScope = {
-  flowid: string;
-  runid: string;
-};
