--- packages/types/src/events/shared/step-event.ts
@@ -1,11 +1,0 @@
-// // import type { PipewarpContext } from "./pipewarp-context.js";
-// // import type { StepEventType } from "../event-map.js";
-// import { CloudEvent } from "./cloud-event.js";
-
-// interface StepScope {
-//   flowid: string;
-//   runid: string;
-//   stepid: string;
-// }
-
-// // type StepEvent<T extends StepEventType> = CloudEvent<T> & StepScope;
