--- packages/types/src/events/shared/worker-event.ts
@@ -1,4 +1,0 @@
-// import type { CloudEvent } from "./cloud-event.js";
-// import type { WorkerEventType } from "./../event-map.js";
-
-// export type WorkerEvent<T extends WorkerEventType> = CloudEvent<T>;
