--- packages/types/src/events/step/action/completed.ts
@@ -1,10 +1,0 @@
-// import type { StepEvent } from "../../shared/step-event.js";
-
-// export type StepActionCompletedData = {
-//   ok: boolean;
-//   message: string;
-//   result?: unknown;
-//   error?: string;
-// };
-
-// type StepActionCompleted = StepEvent<"step.action.completed">;
