+++ packages/types/src/events/step/map.ts
@@ -1,4 +1,4 @@
-import type { DomainActionDescriptor } from "../event-map.js";
+import type { DomainActionDescriptor } from "../shared/otel-attributes.js";
 import { StepCompletedData, StepFailedData, StepStartedData } from "./data.js";
 
 export type StepEventMap = {
