--- packages/types/src/events/worker/registered.ts
@@ -1,11 +1,0 @@
-// import type { WorkerEvent } from "../shared/worker-event.js";
-
-// export type WorkerRegisteredData = {
-//   workerId: string;
-//   status: string;
-//   registeredAt: string;
-// };
-
-// export type WorkerRegistered = WorkerEvent<"worker.registered"> & {
-//   operation: "registered";
-// };
