--- packages/types/src/events/worker/registration-requested.ts
@@ -1,10 +1,0 @@
-// import type { WorkerEvent } from "../shared/worker-event.js";
-// import type { WorkerMetadata } from "../../worker.types.js";
-
-// export type WorkerRegistrationRequestedData = WorkerMetadata;
-
-// export type WorkerRegistrationRequested =
-//   WorkerEvent<"worker.registration.requested"> & {
-//     capability: "registration";
-//     operation: "requested";
-//   };
