+++ packages/types/src/index.ts
@@ -1,17 +1,15 @@
 export * from "./events/shared/index.js";
-// export * from "./events/step/action/completed.js";
-// export * from "./events/step/action/queued.js";
-// export * from "./events/step/mcp/queued.js";
+
 export * from "./events/flow/data.js";
-export * from "./events/engine/data.js";
-export * from "./events/engine/scope.js";
 export * from "./worker.types.js";
 
 export * from "./events/any-event.js";
 export * from "./events/event-map.js";
-// each event type types
+
+// each event types; some are indexed, unsure what form to use;
 export * from "./events/engine/index.js";
 export * from "./events/run/index.js";
+export * from "./events/flow/index.js";
 
 export * from "./events/step/map.js";
 export * from "./events/step/event.js";
@@ -28,5 +26,3 @@ export * from "./events/tool/map.js";
 export * from "./events/worker/data.js";
 export * from "./events/worker/event.js";
 export * from "./events/worker/map.js";
-
-export * from "./events/event-map.js";
